fix: merge 후 에러 수정
Config 중복 파일 삭제
