+++ backendProject/.env.example
@@ -1,0 +1,14 @@
+# 로컬, 운영 환경 관리
+PROFILE=~~~
+
+# RDS 연동
+DB_SERVER=db_server_name
+DB_PORT=db_port
+DB_NAME=db_name
+DB_USERNAME=db_username
+DB_PASSWORD=db_password
+
+REDIS_HOST=redis_server_name
+REDIS_PORT=redis_port
+
+ES_URI=es_uri
