+++ backendProject/build.gradle
@@ -64,14 +64,11 @@ dependencies {
 	// WebClient
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
 
-
-
 	/* env 파일 관련 */
 	implementation 'me.paulschwarz:spring-dotenv:4.0.0'
 
 	/* Google Oauth */
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
-
 
 	/* jwt */
 	implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
