+++ backendProject/src/main/java/likelion/mlb/backendProject/domain/match/service/DraftTimingService.java
@@ -31,8 +31,8 @@ public class DraftTimingService {
         LocalDateTime startedAtKST = r.getStartedAt().atZoneSameInstant(KST).toLocalDateTime();
 
         LocalDate draftDay = startedAtKST.toLocalDate().minusDays(2);
-        LocalDateTime openAt = LocalDateTime.of(draftDay, LocalTime.of(10, 30));
-        LocalDateTime lockAt = LocalDateTime.of(draftDay, LocalTime.of(12, 30));
+        LocalDateTime openAt = LocalDateTime.of(draftDay, LocalTime.of(10, 00));
+        LocalDateTime lockAt = LocalDateTime.of(draftDay, LocalTime.of(18, 30));
 
         return new RoundInfo(
                 r.getId(),
