+++ backendProject/src/main/resources/application.properties
@@ -36,3 +36,17 @@ spring.data.redis.port=${redis.port}
 
 # Elasticsearch
 spring.elasticsearch.uris=${es.uri}
+
+#GoogleOauth
+spring.security.oauth2.client.registration.google.client-id=${GOOGLE_ID}
+spring.security.oauth2.client.registration.google.client-secret=${GOOGLE_SECRET}
+spring.security.oauth2.client.registration.google.redirect-uri=${GOOGLE_REDIRECT}
+spring.security.oauth2.client.registration.google.scope=${GOOGLE_SCOPE}
+
+spring.security.oauth2.client.provider.google.authorization-uri=${GOOGLE_AUTHORIZATION}
+spring.security.oauth2.client.provider.google.token-uri=${GOOGLE_TOKEN}
+spring.security.oauth2.client.provider.google.user-info-uri=${GOOGLE_USER_INFO}
+spring.security.oauth2.client.provider.google.user-name-attribute=${GOOGLE_USER_NAME_ATTRIBUTE}
+
+#jwt
+jwt.secret = ${JWT_SECRET}
